Suppl 8 + WIP cleverers suppl
There was bug in merge_annovar_cnok.R: header=TRUE.
We always lost the first mutation, but reasonably this did not create any issues because we have repeated
info in all.annovar.gz for clones due to the repeated input of all_pairs_annovar - this is then fixed
doing pivot tables or uniq so we should be ok. Will rerun all edt2/8 and check to be sure.